# Conflicts:
#   src/main/java/com/ra/base_spring_boot/dto/response/CenterResponseDTO.java
#   src/main/java/com/ra/base_spring_boot/service/impl/CenterServiceImpl.java

+++ src/main/java/com/ra/base_spring_boot/service/impl/PartnerServiceImpl.java
@@ -51,7 +51,7 @@ public class PartnerServiceImpl implements IPartnerService {
     @Override
     @Transactional
     public PartnerResponseDTO updatePartner(Long id, PartnerDTO dto) {
-        Partner partner = partnerRepository.findById(Math.toIntExact(id))
+        Partner partner = partnerRepository.findById(id)
                 .orElseThrow(() -> new UsernameNotFoundException("Không tìm thấy đối tác"));
 
         if (!partner.getPartnerCode().equals(dto.getPartnerCode()) &&
@@ -73,7 +73,7 @@ public class PartnerServiceImpl implements IPartnerService {
 
     @Override
     public PartnerResponseDTO getPartnerById(Long id) {
-        Partner partner = partnerRepository.findById(Math.toIntExact(id))
+        Partner partner = partnerRepository.findById(id)
                 .orElseThrow(() -> new UsernameNotFoundException("Không tìm thấy đối tác"));
         return mapEntityToResponse(partner);
     }
@@ -88,7 +88,7 @@ public class PartnerServiceImpl implements IPartnerService {
 
     @Override
     public void deletePartner(Long id) {
-        Partner partner = partnerRepository.findById(Math.toIntExact(id))
+        Partner partner = partnerRepository.findById(id)
                 .orElseThrow(() -> new UsernameNotFoundException("Không tìm thấy đối tác"));
 
         if (partner.getAvatarUrl() != null) {
@@ -130,14 +130,14 @@ public class PartnerServiceImpl implements IPartnerService {
 
     @Override
     @Transactional(readOnly = true)
-    public GetDetailPartnerResponse getDetailPartner(int partnerId) {
+    public GetDetailPartnerResponse getDetailPartner(Long partnerId) {
         Partner partner = partnerRepository.findById(partnerId)
                 .orElseThrow(() -> new UsernameNotFoundException("Không tìm thấy đối tác"));
 
         PartnerResponseDTO partnerDTO = mapEntityToResponse(partner);
 
         List<Course> courses = courseRepository.findByPartnerId(partnerId);
-        List<CourseOff> courseOffs = courseOffRepository.findByPartnerId((long) partnerId);
+        List<CourseOff> courseOffs = courseOffRepository.findByPartnerId(partnerId);
 
         List<Long> courseIds = courses.stream().map(Course::getId).toList();
         List<Long> courseOffIds = courseOffs.stream().map(CourseOff::getId).toList();
